requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
APScheduler==3.11.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
discord.py==2.5.2
feedparser==6.0.11
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
propcache==0.3.2
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.1
requests==2.32.4
sgmllib3k==1.0.0
tzlocal==5.3.1
urllib3==2.5.0
yarl==1.20.1
